+++ client/my-sites/site-settings/settings-writing/controller.js
@@ -7,7 +7,7 @@ import React from 'react';
  * Internal dependencies
  */
 import { renderWithReduxStore } from 'lib/react-helpers';
-import WritingMain from 'my-sites/site-settings/writing/main';
+import WritingMain from 'my-sites/site-settings/settings-writing/main';
 import Taxonomies from 'my-sites/site-settings/taxonomies';
 
 export default {
